+++ R/get-cur-asc-qr-data.R
@@ -1,4 +1,4 @@
-#' Get Current Ambulatory Surgery Center Data.
+#' Get Current Ambulatory Surgery Center Quality Reporting Data.
 #'
 #' @family Hospital Data
 #'
@@ -14,15 +14,15 @@
 #' of those data sets you would like,
 #'
 #' @param .data The data that results from the `current_hosp_data()` function.
-#' @param .data_sets The default is: c("Facility","State","National"), which will
-#' bring back all of the data in the Ambulatory Surgery Center data sets that
-#' are in the file. You can choose from the following:
+#' @param .data_sets The default is: c("Facility","State","National"), which
+#' will bring back all of the data sets that are in the Ambulatory Surgery Center
+#' OAS CAHPS data sets. You can choose from the following:
 #' *  Facility
 #' *  National
 #' *  State
 #'
-#' You can also pass things like c("state","Nation") as behind the scense only
-#' the Ambulatory Surgery Center datasets are available to the function to choose
+#' You can also pass things like c("state","Nation") as behind the scenes only
+#' the Ambulatory Surgery Center data sets are available to the function to choose
 #' from and `grep` is used to find matches with `ignore.case = TRUE` set.
 #'
 #' @examples
@@ -30,20 +30,20 @@
 #' library(dplyr)
 #'
 #' current_hosp_data() %>%
-#'   current_asc_data(.data_sets = c("State","National"))
+#'   current_asc_oas_cahps_data(.data_sets = c("State","National"))
 #' }
 #'
 #' @return
 #' Gets the current ASC data from the current hospital data file.
 #'
-#' @name current_asc_data
+#' @name current_asc_oas_cahps_data
 NULL
 
 #' @export
-#' @rdname current_asc_data
+#' @rdname current_asc_oas_cahps_data
 
-current_asc_data <- function(.data,
-                             .data_sets = c("Facility","State","National")) {
+current_asc_oas_cahps_data <- function(.data,
+                                       .data_sets = c("Facility","State","National")) {
 
     # Variables
     ds <- .data_sets
@@ -59,8 +59,12 @@ current_asc_data <- function(.data,
 
     # Manipulations
     # Get the exact files necessary to the ASC
-    file_names_vec <- c("ASC_Facility.csv","ASC_National.csv","ASC_State.csv")
+    file_names_vec <- c("ASCQR_OAS_CAHPS_BY_ASC.csv",
+                        "ASCQR_OAS_CAHPS_NATIONAL.csv",
+                        "ASCQR_OAS_CAHPS_STATE.csv")
+
     asc_list <- l[names(l) %in% file_names_vec]
+    names(asc_list)[1] <-  "ASCQR_OAS_CAHPS_FACILITY.csv"
 
     # Make sure there are no 0 length items
     asc_list <- asc_list[lengths(asc_list) > 0]
@@ -70,11 +74,11 @@ current_asc_data <- function(.data,
         paste(ds, collapse = "|"),
         names(asc_list),
         ignore.case = TRUE
-        )]
+    )]
 
     # Return\
-    attr(ret, ".list_type") <- "current_asc_list"
-    class(ret) <- c("current_asc_list", class(ret))
+    attr(ret, ".list_type") <- "current_asc_oascahps_list"
+    class(ret) <- c("current_asc_oascahps_list", class(ret))
 
     return(ret)